Display videos on the landing page correctly.
Previously, videos were rendered correctly in a post, but they were not
rendered correctly on the landing page or potentially also not on
another page that displayed various postings in a query loop.

The reason for this was that the JavaScript and CSS stylesheets were
not included by such pages.

This commit makes sure that those elements are properly included
whenever a video is included in a page.